automate st3 package install

+++ init-dots.sh
@@ -51,8 +51,10 @@ for PACKAGE in $SOURCE_DIR/*; do
     [ ! -d "$TARGET_DIR/$PACKAGE" ] && mkdir -p "$TARGET_DIR/$PACKAGE"
     for FILE in $SOURCE_DIR/$PACKAGE/*; do
         FILE=$(basename "$FILE")
-        [ -d "$TARGET_DIR/$PACKAGE/$FILE" ] && rm -rf "${TARGET_DIR:?}/$PACKAGE/$FILE"
-        ln -nsf "$SOURCE_DIR/$PACKAGE/$FILE" "$TARGET_DIR/$PACKAGE/$FILE"
+        if [ "$FILE" != "Package Control.sublime-package" ]; then
+            [ -d "$TARGET_DIR/$PACKAGE/$FILE" ] && rm -rf "${TARGET_DIR:?}/$PACKAGE/$FILE"
+            ln -nsf "$SOURCE_DIR/$PACKAGE/$FILE" "$TARGET_DIR/$PACKAGE/$FILE"
+        fi
     done
 done
